Refactor Booking and Livewire forms to utilize new Filament actions and schemas, enhancing bulk action capabilities and updating form structure for better maintainability.
